home

+++ blog/article.php
@@ -1,4 +1,8 @@
- <!doctype html>
+<?php 
+	error_reporting(E_ALL);
+?>
+
+<!doctype html>
 <html>
 	<head>
 		<meta charset="utf-8">
@@ -8,11 +12,19 @@
 	
 		<?php
 			$name = $_GET['id'];
-			$text = file_get_contents("data/$name");
-			
-
-			echo "<h1>$name</h1>";
-			echo "<div>$text</div>";
+		
+			// file_exists -- Проверить наличие указанного файла или каталога
+			if($name != '' && file_exists("data/$name") && is_file("data/$name")) {
+				$text = file_get_contents("data/$name");
+				echo "<h1>$name</h1>";
+				echo "<div>$text</div>";
+			} else {
+				echo '
+					<h1>404</h1>
+					<hr>
+					<a href="post.php">post</a><br>
+				';
+			}
 		?>
 		<hr>
 		<a href="post.php">post</a><br>
